[feat] Make stopwatch
Made a stopwatch for tracking how much time it takes for player to solve the puzzle. Stopwatch starts when player do first touch on the slime and stops when player wins. Stopwatch shows hours, minutes and seconds. Also it remembers the minimum solve time for each puzzle mode, but doesn't display it for now. Added custom font in text mesh pro for stopwatch stats.