[feat] Scene transition
Made a smooth scenes transition that makes screen dark from borders to center until whole screen is black and then just do the same but in reverse.
Transition effect was made by changing the radius of the shader from TransitionManager. This shader is applied to material that applied to Image that exist in every scene and that image has special tag to make it findable for TransitionManager.
Also I changed canvas plane distance to 1 to Transition was displayer above all UI and game models.
And the last thing is I changed variable name in Stopwatch script.