[fix] Unsolvable puzzle problem
Before puzzle was unsolvable sometimes because grid was spawned randomly. So I remade algorithm and now it creates solved puzzle and then shuffle grid. Of course because of this strategy grid can be shuffled strongly or weakly, but anyway it is better then just have unsolvable puzzle.
Also did small fix in Victory Menu script to avoid errors in consol.